+++ app/AssertionRemoval.hs
@@ -8,24 +8,17 @@ import Control.Monad (foldM, (=<<), (<=<), (>>=))
 
 import Data.Map (Map, (!))
 import qualified Data.Map as Map
-import Data.List (find)
+import Data.List (find, nub)
 import Data.Maybe
 
 import Syntax
 import AstReversing
+import EvalExpr
 
 type Vars = Map Ident AST
 
 intSz = 32
 
-fst3 :: (a, b, c) -> a
-fst3 (x, _, _) = x
-
-snd3 :: (a, b, c) -> b
-snd3 (_, x, _) = x
-
-trd3 :: (a, b, c) -> c
-trd3 (_, _, x) = x
 
 makeVar :: Type -> Ident -> Z3 AST
 makeVar t n =
@@ -47,32 +40,43 @@ makeArr t n =
       Z3.mkFreshConst (show n) arrSort
 
 
-setArray :: Vars -> [Expr] -> Int -> AST -> Z3 AST
+-- If Nothing is returned the array being set should be invalidated
+setArray :: Vars -> [Expr] -> Int -> AST -> Z3 (Either Ident AST)
 setArray scope vals idx arr =
   case vals of
-    [] -> return arr
+    [] -> return $ Right arr
     (hd:tl) -> do
       idx' <- Z3.mkFreshBvVar "arrIdx" intSz
       hd'  <- processExpr scope hd
-      Z3.mkStore arr idx' hd' >>= (setArray scope tl (idx+1))
+      case hd' of
+        Right x ->
+          Z3.mkStore arr idx' x >>= (setArray scope tl (idx+1))
+        otherwise -> return otherwise
 
--- Used during renaming, as not all variables need to be renamed
+
+
+-- If this reports Nothing during processing it's because name is invalidated.
+-- If it happens during renaming, it's because not all variables need to be renamed
 tryGetVar :: Ident -> Map Ident a -> Maybe a
 tryGetVar name scope =
   Map.lookup name scope
 
-getVar :: Ident -> Map Ident a -> String -> a
-getVar n scope errorCode =
-  case Map.lookup n scope of
-    Just x -> x
-    Nothing -> error $ "Variable " ++ show n ++ " could not be found! " ++ errorCode
-      ++ foldl (\a b -> a ++ ", " ++ show b) "" (Map.keys scope)
 
--- getVar :: Ident -> Vars -> AST
--- getVar n scope =
---   case Map.lookup n scope of
---     Just x -> x
---     Nothing -> error $ "Variable " ++ show n ++ " could not be found!"
+invalidateVars :: [Ident] -> Vars -> Z3 Vars 
+invalidateVars stmts scope =
+  foldM (\s key -> return $ Map.delete key s) scope stmts
+
+modifiedVars :: [Stmt] -> [Ident]
+modifiedVars = \case
+  [] -> []
+
+  (Mod (Moderator (Var _ name _ _) _ _) _ _:tl)   -> name : modifiedVars tl
+  (Mod (Moderator (Arr _ name _ _ _) _ _) _ _:tl) -> name : modifiedVars tl
+
+  (Switch (Var _ name1 _ _) (Var _ name2 _ _) _:tl) -> name1 : name2 : modifiedVars tl
+  (Switch (Arr _ name1 _ _ _) (Arr _ name2 _ _ _) _:tl) -> name1 : name2 : modifiedVars tl
+
+  (_:tl) -> modifiedVars tl
 
 
 operate :: BinOp -> AST -> AST -> Z3 AST
@@ -98,45 +102,75 @@ operate opr e1 e2 =
     Eq     -> Z3.mkEq e1 e2
 
 
-processExpr :: Vars -> Expr -> Z3 AST
+-- If this reports Nothing the variable using this expression should be invalidated
+processExpr :: Vars -> Expr -> Z3 (Either Ident AST)
 processExpr scope e =
   case e of
     ConstE v ->
       case v of
-        IntegerV i -> Z3.mkBvNum intSz i
-        BooleanV b -> Z3.mkBool b
+        IntegerV i -> Z3.mkBvNum intSz i >>= (\val -> return $ Right val)
+        BooleanV b -> Z3.mkBool b >>= (\val -> return $ Right val)
+
     VarE var ->
       case var of
         LVar n ->
-          return $ getVar n scope "While processing expression"
+          case tryGetVar n scope of
+            Just n' -> return $ Right n'
+            Nothing -> return (Left n)
+
         _ -> error "VarE for array not implemented yet"
-    Arith op e1 e2 -> do -- (+ (- 2 c) l)
+
+    Arith op e1 e2 -> do
       e1' <- processExpr scope e1
       e2' <- processExpr scope e2
-      operate op e1' e2'
+      case (e1', e2') of
+        (Right x, Right y) ->
+          operate op x y >>= (\val -> return $ Right val)
+        (Left n, _) ->
+          return (Left n)
+        (_, Left n) ->
+          return (Left n)
+
     Not e -> do
       e' <- processExpr scope e
-      Z3.mkNot e'
+      case e' of
+        Right x -> 
+          Z3.mkNot x >>= (\val -> return $ Right val)
+        
+        Left n -> return (Left n)
     _ -> error $ "Optimization expression: Expr not implemented yet " ++ show e
 
 
 createITE :: AST -> Vars -> Vars -> Vars -> Z3 Vars
 createITE cond orig scope1 scope2 = foldM f orig $ Map.keys orig
   where f scope var = do
-          let var1  = getVar var scope1 "While calling createITE"
-          let var2  = getVar var scope2 "While calling createITE"
+          -- During this phase it should always be Just, as we work on the origin scope
+          let var1   = tryGetVar var scope1
+          let var2   = tryGetVar var scope2
+
           newVar    <- makeVar (IntegerT) var -- TODO: again a hack with the type
-          ite       <- Z3.mkIte cond var1 var2
+          ite       <- Z3.mkIte cond (fromJust var1) (fromJust var2)
           Z3.assert =<< Z3.mkEq newVar ite
           return $ Map.insert var newVar scope
 
 
-validate :: Z3 Bool
-validate = do
+validate :: Stmt -> AST -> Z3 (Either String Bool) --(Bool, String)
+validate stmt expr = do
+  Z3.push
+  Z3.assert expr
   val <- Z3.solverCheck
   case val of
-    Z3.Sat -> return False
-    Z3.Unsat -> return True
+    Z3.Sat -> do
+      Z3.pop 1
+      Z3.push
+      Z3.assert =<< Z3.mkNot expr
+      val' <- Z3.solverCheck
+      Z3.pop 1
+      case val' of
+        Z3.Sat -> return $ Right False
+        Z3.Unsat -> return $ Right True
+    
+    Z3.Unsat -> Z3.pop 1 >> return (Left $ "Z3 expression is a fallacy: " ++ show stmt)
 
 
 -- Processes stmt in given ast and scope, and returns (updated scope, optimized stmt)
@@ -144,164 +178,378 @@ validate = do
 --    * dealloc is met
 --    * ite is met
 --    * for loop is met
-processStatement :: Bool -> [ProcDecl] -> Vars -> Stmt -> Int -> Z3 (Vars, Stmt)
-processStatement doOpt ast scope stmt state =
+processStatement :: Bool -> [ProcDecl] -> Vars -> Stmt -> Int -> [String] -> Z3 (Vars, Stmt, [String])
+processStatement doOpt ast scope stmt state warnings =
   case stmt of
     Global (Var t name val _) _ -> do
       newVar <- makeVar (fromJust t) name
-      newVal <- processExpr scope (fromJust val)
-      Z3.assert =<< Z3.mkEq newVar newVal
-      return (Map.insert name newVar scope, stmt)
+
+      processExpr scope (fromJust val) >>= \case
+        Right newVal -> do
+          Z3.assert =<< Z3.mkEq newVar newVal
+          return (Map.insert name newVar scope, stmt, warnings)
+
+        Left errvar ->
+          return (scope, stmt, (show name ++ " used with invalid variable " ++ show errvar) : warnings)
 
     Global (Arr t name sz vals _) _ -> do
       newVar  <- makeArr (fromJust t) name
-      newVar' <- setArray scope (fromJust vals) 0 newVar
-      return (Map.insert name newVar' scope, stmt)
+      setArray scope (fromJust vals) 0 newVar >>= \case
+        Right newVar' ->
+          return (Map.insert name newVar' scope, stmt, warnings)
+
+        Left errvar ->
+          return (scope, stmt, (show name ++ " used with invalid variable " ++ show errvar) : warnings)
 
     Local (Var t name val _) _ -> do
       newVar <- makeVar (fromJust t) name
-      newVal <- processExpr scope (fromJust val)
-      Z3.assert =<< Z3.mkEq newVar newVal
-      return (Map.insert name newVar scope, stmt)
 
-    DLocal (Var t name val _) _ -> do
-      if (doOpt) then do
-        val     <- processExpr scope (fromJust val)
-        let var = getVar name scope "While processing dealloc"
-        eq <- Z3.mkEq var val
+      processExpr scope (fromJust val) >>= \case
+        Right newVal -> do
+          Z3.assert =<< Z3.mkEq newVar newVal
+          return (Map.insert name newVar scope, stmt, warnings)
+
+        Left errvar ->
+          return (scope, stmt, (show name ++ " used with invalid variable " ++ show errvar) : warnings)
+
+    DLocal (Var t name val _) _ ->
+      if (doOpt) then
+        -- val     <- processExpr scope (fromJust val)
+        processExpr scope (fromJust val) >>= \case
+          Right val ->
+            case tryGetVar name scope of
+              Just var -> do
+                eq <- Z3.mkEq var val
+                validated <- validate stmt eq
+                case validated of
+                  Left err    ->
+                    -- Continue without the assertion having any effect on the analysis
+                    return (scope, stmt, err:warnings)
+                  Right True  -> return (scope, Skip, warnings)
+                  Right False -> Z3.assert eq >> return (scope, stmt, warnings)
+
+              Nothing ->
+                -- The deallocated variable is invalidated so the information cannot be used further on
+                return (scope, stmt, (show name ++ " has been invalidated") : warnings)
+
+          Left errvar ->
+            return (scope, stmt, (show name ++ " used with invalid variable " ++ show errvar) : warnings)
       
-        Z3.push
-        Z3.assert =<< Z3.mkNot eq
-        validated <- validate
-        if (validated) then Z3.pop 1 >> return (scope, Skip)
-        else                Z3.pop 1 >> Z3.assert eq >> return (scope, stmt)
+        
       else
-        return (scope, stmt)
+        return (scope, stmt, warnings)
 
-    Mod (Moderator (Var _ name _ _) op val) _ _ -> do
+    Mod (Moderator (Var _ name _ _) op val) _ _ ->
       -- Moderation can only happen for integer types. This should be checked by a type checker
-      let oldVar  = getVar name scope "While processing moderator for var"
-      newVar      <- makeVar IntegerT name
-      incrVal     <- processExpr scope val
-      newVal      <-
-        case op of
-          PlusEq -> Z3.mkBvadd oldVar incrVal
-          SubEq  -> Z3.mkBvsub oldVar incrVal
-          XorEq  -> Z3.mkBvxor oldVar incrVal
-      Z3.assert =<< Z3.mkEq newVar newVal
-      return (Map.insert name newVar scope, stmt)
+      mod stmt scope warnings
 
-    Mod (Moderator (Arr _ name sz vals _) op val) idx _ -> do
-      let arr = getVar name scope "While processing moderator for array"
-      incrVal <- processExpr scope val
-      idx'    <- processExpr scope (head (fromJust idx)) -- TODO: Only works for 1D array
-      oldVal  <- Z3.mkSelect arr idx'
-      newVal  <-
-        case op of
-          PlusEq -> Z3.mkBvadd oldVal incrVal
-          SubEq  -> Z3.mkBvsub oldVal incrVal
-          XorEq  -> Z3.mkBvxor oldVal incrVal
-      Z3.mkStore arr idx' newVal
-      return (scope, stmt)
+    Mod (Moderator (Arr _ name sz vals _) op val) idx _ ->
+      case tryGetVar name scope of
+        Just arr ->
+          processExpr scope val >>= \case
+            Right incrVal ->
+              -- TODO: Only works for 1D arrays
+              processExpr scope (head (fromJust idx)) >>= \case
+                Right idx' -> do
+                  oldVal  <- Z3.mkSelect arr idx'
+                  newVal  <-
+                    case op of
+                      PlusEq -> Z3.mkBvadd oldVal incrVal
+                      SubEq  -> Z3.mkBvsub oldVal incrVal
+                      XorEq  -> Z3.mkBvxor oldVal incrVal
+                  Z3.mkStore arr idx' newVal
+                  return (scope, stmt, warnings)
+                Left errvar ->
+                  return (Map.delete name scope, stmt, (show name ++ " just got invalidated because of " ++ show errvar) : warnings)
+            Left errvar ->
+              return (Map.delete name scope, stmt, (show name ++ " just got invalidated because of " ++ show errvar) : warnings)
+        Nothing ->
+          return (scope, stmt, (show name ++ " has been invalidated") : warnings)
 
     Switch (Var t1 n1 _ _) (Var t2 n2 _ _) _ -> do
-      newVar1 <- makeVar (fromJust t1) n1 -- TODO: This requires a type check phase to annotate with types
-      newVar2 <- makeVar (fromJust t2) n2
-      let n1' =  getVar n1 scope "While processing switch for variables"
-      let n2' =  getVar n2 scope "While processing switch for variables"
-      Z3.assert =<< Z3.mkEq newVar1 n2'
-      Z3.assert =<< Z3.mkEq newVar2 n1'
-      return (Map.insert n1 newVar1 $ Map.insert n2 newVar2 scope, stmt)
+      -- newVar1 <- makeVar (fromJust t1) n1
+      -- newVar2 <- makeVar (fromJust t2) n2
+      newVar1 <- makeVar IntegerT n1 -- TODO: Stop this hardcoding of types
+      newVar2 <- makeVar IntegerT n2 -- TODO: Stop this hardcoding of types
+
+      case (tryGetVar n1 scope, tryGetVar n2 scope) of
+        (Just n1', Just n2') -> do
+          Z3.assert =<< Z3.mkEq newVar1 n2'
+          Z3.assert =<< Z3.mkEq newVar2 n1'
+          return (Map.insert n1 newVar1 $ Map.insert n2 newVar2 scope, stmt, warnings)
+
+        _ ->
+          -- One of the switched variables has been invalidated
+          return (scope, stmt, ("Either " ++ show n1 ++ " or " ++ show n2 ++ " has been invalidated") : warnings)
 
     Switch (Arr t1 n1 _ _ _) (Arr t2 n2 _ _ _) _ -> do
-      newVar1 <- makeVar (fromJust t1) n1 -- TODO: This requires a type check phase to annotate with types
-      newVar2 <- makeVar (fromJust t2) n2
-      let n1' =  getVar n1 scope "While processing switch for arrays"
-      let n2' =  getVar n2 scope "While processing switch for arrays"
-      Z3.assert =<< Z3.mkEq newVar1 n2'
-      Z3.assert =<< Z3.mkEq newVar2 n1'
-      return (Map.insert n1 newVar1 $ Map.insert n2 newVar2 scope, stmt)
+      -- newVar1 <- makeVar (fromJust t1) n1
+      -- newVar2 <- makeVar (fromJust t2) n2
+      newVar1 <- makeVar IntegerT n1 -- TODO: Stop this hardcoding of types
+      newVar2 <- makeVar IntegerT n2 -- TODO: Stop this hardcoding of types
 
-    Ite ifcond body1 ficond body2 pos -> do
+      case (tryGetVar n1 scope, tryGetVar n2 scope) of
+        (Just n1', Just n2') -> do
+          Z3.assert =<< Z3.mkEq newVar1 n2'
+          Z3.assert =<< Z3.mkEq newVar2 n1'
+          return (Map.insert n1 newVar1 $ Map.insert n2 newVar2 scope, stmt, warnings)
+
+        _ ->
+          return (scope, stmt, ("Either " ++ show n1 ++ " or " ++ show n2 ++ " has been invalidated") : warnings)
+
+    Ite ifcond body1 ficond body2 pos ->
       {-
         1. validate if body
         2. validate fi cond
         3. validate else body
         4. make ITE continuation
       -}
-      ifcond' <- processExpr scope ifcond
-      if (doOpt) then do
-        -- if part
-        Z3.push
-        Z3.assert ifcond'
-        body1'  <- processStatements doOpt ast scope body1
-        ficond' <- processExpr (fst3 body1') ficond
+      processExpr scope ifcond >>= \case
+        Right ifcond' ->
+          if (doOpt) then do
+            -- if part
+            Z3.push
+            Z3.assert ifcond'
+            -- body1'  <- processStatements doOpt ast scope body1 0
+            (_, scope', _, _) <- processStatements body1 doOpt ast scope 0 []
+            processExpr scope' ficond >>= \case
+              Right ficond' -> do
+                -- fi part
+                fiValidated <- validate stmt ficond'
+                Z3.pop 1
 
-        -- fi part
-        Z3.assert =<< Z3.mkNot ficond'
-        fiValidated <- validate
-        Z3.pop 1
+                -- else part
+                Z3.push
+                Z3.assert =<< Z3.mkNot ifcond'
+                -- processStatements doOpt ast scope body2 0
+                processStatements body2 doOpt ast scope 0 []
+                Z3.pop 1
 
-        -- else part
-        Z3.push
-        Z3.assert =<< Z3.mkNot ifcond'
-        processStatements doOpt ast scope body2
-        Z3.pop 1
+                -- Need to create the scope of body1 and body2 again, because pop erased them
+                -- TODO: Making change whole push pop above, to make sure we don't need this
+                -- body1'  <- processStatements doOpt ast scope body1 0
+                -- body2'  <- processStatements doOpt ast scope body2 0
+                (body1', scope1, state1, warnings') <- processStatements body1 doOpt ast scope 0 warnings
+                (body2', scope2, state2, warnings'') <- processStatements body2 doOpt ast scope 0 warnings'
 
-        -- Need to create the scope of body1 and body2 again, because pop erased them
-        -- TODO: Making change whole push pop above, to make sure we don't need this
-        body1'  <- processStatements doOpt ast scope body1
-        body2'  <- processStatements doOpt ast scope body2
+                -- Continuation
+                ite <- createITE ifcond' scope scope1 scope2
 
-        -- Continuation
-        ite <- createITE ifcond' scope (fst3 body1') (fst3 body2')
-        return (ite,
-          Ite ifcond (snd3 body1') (if fiValidated then SkipE else ficond) (snd3 body2') pos)
-      else do
-        body1'  <- processStatements doOpt ast scope body1
-        body2'  <- processStatements doOpt ast scope body2
-        ite <- createITE ifcond' scope (fst3 body1') (fst3 body2')
-        return (ite, stmt)
+                case fiValidated of
+                  Left err    ->
+                    return (ite, Ite ifcond body1' ficond body2' pos, err:warnings'')
+                  Right True  ->
+                    return (ite, Ite ifcond body1' SkipE body2' pos, warnings'')
+                  Right False ->
+                    return (ite, Ite ifcond body1' ficond body2' pos, warnings'')
 
-    Assert e _ -> do
-      cond <- processExpr scope e
+              Left errvar -> do
+                -- Cannot guarantee anything as the fi conditional uses invalidated variables
+                -- Must invalidate any variables modified within both bodies
+                scope' <- invalidateVars (nub $ modifiedVars body1 ++ modifiedVars body2) scope
+                return (scope', stmt, (show errvar ++ " invalidated if-else statement") : warnings)
+          else do
+            -- body1'  <- processStatements doOpt ast scope body1 0
+            -- body2'  <- processStatements doOpt ast scope body2 0
+            (body1', scope1, state1, warnings') <- processStatements body1 doOpt ast scope 0 warnings
+            (body2', scope2, state2, warnings'') <- processStatements body2 doOpt ast scope 0 warnings'
+            ite <- createITE ifcond' scope scope1 scope2
+            return (ite, stmt, warnings'')
 
-      Z3.push
-      Z3.assert =<< Z3.mkNot cond
-      validated <- validate
-      Z3.pop 1
+        Left errvar -> do
+          -- Cannot guarantee anything as the if conditional uses invalidated variables
+          -- Must invalidate any variables modified within both bodies
+          scope' <- invalidateVars (nub $ modifiedVars body1 ++ modifiedVars body2) scope
+          return (scope', stmt, (show errvar ++ " invalidated if-else statement") : warnings)
 
-      Z3.assert cond >> return (scope, stmt)
 
+    Assert e _ ->
+      processExpr scope e >>= \case
+        Right cond -> do
+          validated <- validate stmt cond
+          case validated of
+            Left err    ->
+              return (scope, stmt, err:warnings)
+            Right True  ->
+              return (scope, Skip, warnings)
+            Right False ->
+              Z3.assert cond >> return (scope, stmt, warnings)
+
+        Left errvar ->
+          return (scope, stmt, (show stmt ++ " used invalidated var") : warnings)
+
+
+      
+
+    {-
+      1. Rename formal parameters and local variables in called procedure.
+      2. Connect FArg with AArg.
+      3. Process the damn body,
+      4. Connect parameters back again
+    -}
     Call name aargs pos ->
-      {-
-          1. Rename formal parameters and local variables in called procedure.
-          2. Connect FArg with AArg.
-          3. Process the damn body,
-          4. Connect parameters back again
-      -}
-      processFunctionCall scope aargs (findProc name ast) state
+      processFunctionCall scope aargs (findProc name ast) state warnings
 
     Uncall name aargs pos ->
       let ast' = reverseProcedure (findProc name ast) [] in
-        processFunctionCall scope aargs (head ast') state
+        processFunctionCall scope aargs (head ast') state warnings
 
 
-    For1 var body mod cond pos -> error "Optimization statement: For1 not implemented yet"
-    For2 var mod body cond pos -> error "Optimization statement: For2 not implemented yet"
+    {-
+      1. Check if it has a loop invariant.
+      2. if yes then opt with this else check for const for
+      3. if yes then opt with unroll else stop
 
-    Skip -> return (scope, Skip)
+      Loop invariant must be true at entry and exit of each iteration.
+
+      Rules for unrolling:
+        No variables used at:
+          * Instantiation,
+          * Increment,
+          * End condition.
+    -}
+    For1 var body m cond b pos -> do
+      (scope', b', body', warnings') <- processForLoop True var body m cond pos warnings
+      return (scope', For1 var body' m cond b' pos, warnings')
+
+    For2 var m body cond b pos -> do
+      (scope', b', body', warnings') <- processForLoop False var body m cond pos warnings
+      return (scope', For2 var m body' cond b' pos, warnings')
+
+    Skip -> return (scope, Skip, warnings)
     _ -> error $ "Optimization statement: Stmt not implemented yet - " ++ show stmt
   where
-    processFunctionCall :: Vars -> [AArg] -> ProcDecl -> Int -> Z3 (Vars, Stmt)
-    processFunctionCall scope aargs p state =  do
+    mod :: Stmt -> Vars -> [String] -> Z3 (Vars, Stmt, [String])
+    mod (Mod (Moderator (Var _ name _ _) op val) _ _) scope warnings = do
+      case tryGetVar name scope of
+        Just oldVar -> do
+          newVar      <- makeVar IntegerT name
+          processExpr scope val >>= \case
+            Right incrVal -> do
+              newVal <-
+                case op of
+                  PlusEq -> Z3.mkBvadd oldVar incrVal
+                  SubEq  -> Z3.mkBvsub oldVar incrVal
+                  XorEq  -> Z3.mkBvxor oldVar incrVal
+              Z3.assert =<< Z3.mkEq newVar newVal
+              return (Map.insert name newVar scope, stmt, warnings)
+
+            Left errvar ->
+              return (Map.delete name scope, stmt, (show name ++ " just got invalidated because of " ++ show errvar) : warnings)
+
+        _ ->
+          return (scope, stmt, (show name ++ " used invalidated var") : warnings)
+    
+    mod _ _ _ = error "Please only use this function for processing modification!"
+
+    -- returns (new scope, outer loop optimizable?, optimized body)
+    processForLoop :: Bool -> Var -> [Stmt] -> Moderator -> Expr -> Pos -> [String] -> Z3 (Vars, Bool, [Stmt], [String])
+    processForLoop forward var@(Var t name (Just val) p1) body m@(Moderator v op incr) cond pos warnings =
+      if (varFreeExprs [val, cond, incr] && modificationFreeIncr name body) then
+        -- Safe to calculate unroll number and unroll
+        case fromJust t of
+          IntegerT -> do
+            let start = evalConstantIntExpr val
+            let end   = evalConstantIntExpr cond
+            let incr'  = evalConstantIntExpr incr
+            case op of
+              XorEq  -> error "xor for loops not implemented yet"
+
+              _ ->
+                if start > end then error "Loops utilizing overflow/underflow is not allowed yet"
+                else do
+                  let bound = (end - start) `div` incr'
+
+                  -- creating the initial variable
+                  z3var  <- makeVar (fromJust t) name
+                  processExpr scope val >>= \case
+                    Right z3expr -> do
+                      eq     <- Z3.mkEq z3var z3expr
+                      Z3.assert eq
+
+                      Z3.push
+                      -- (scope', body', state') <- processStatements True ast (Map.insert name z3var scope) body state
+                      (body', scope', state', warnings') <- processStatements body doOpt ast (Map.insert name z3var scope) state warnings
+                      Z3.pop 1
+
+                      (scope'', validated, warnings'') <- unroll bound forward body' pos var z3expr m scope' state' warnings'
+                      if (validated) then
+                        return (scope'', True, body', warnings'')
+                      else
+                        return (scope'', False, body', warnings'')
+                    Left errvar ->
+                      return (scope, False, body, (show name ++ " used invalid variable") : warnings)
+
+          BooleanT -> error "Boolean for-loops not implemented yet"
+      else --error "Loop not being unrollable not implemented yet"
+      {-
+        Must invalidate every every variable modified within the loop.
+        This is done by removing them from scope.
+      -}
+        invalidateVars (modifiedVars body) scope
+         >>= (\scope' -> return (scope', False, body, ("Loop could not be unrolled: " ++ show pos) : warnings))
+      where
+        varFreeExprs :: [Expr] -> Bool
+        varFreeExprs = \case
+          []            -> True
+          (ConstE _:tl)      -> True && varFreeExprs tl
+          (VarE _:tl)        -> False && varFreeExprs tl
+          (Arith _ e1 e2:tl) -> varFreeExprs [e1, e2] && varFreeExprs tl
+          (Not e:tl)         -> varFreeExprs [e] && varFreeExprs tl
+          (Size n:tl)        -> False && varFreeExprs tl -- TODO: Should be True, if it worked properly
+          (SkipE:tl)         -> True && varFreeExprs tl
+        
+        modificationFreeIncr :: Ident -> [Stmt] -> Bool
+        modificationFreeIncr name1 = \case
+          [] -> True
+          (Mod (Moderator (Var _ name2 _ _) _ _) _ _:tl) ->
+            if (name1 == name2) then False else modificationFreeIncr name1 tl
+          (_:tl) -> modificationFreeIncr name1 tl 
+
+        unroll :: Int -> Bool -> [Stmt] -> Pos -> Var -> AST -> Moderator -> Vars -> Int -> [String] -> Z3 (Vars, Bool, [String])
+        unroll bound forward body pos var@(Var t name _ _) val m scope state warnings =
+          case bound of
+            0 -> return (scope, True, warnings)
+            _ -> do
+              (scope'', state', warnings') <-
+                if (forward) then do
+                  -- (scope', _, state') <- processStatements False ast scope body state
+                  (_, scope', state', _) <- processStatements body False ast scope state []
+                  (scope'', _, warnings') <- mod (Mod m Nothing pos) scope' warnings
+                  return (scope'', state', warnings')
+                else do
+                  (scope', _, warnings') <- mod (Mod m Nothing pos) scope warnings
+                  -- (scope'', _, state') <- processStatements False ast scope' body state
+                  (_, scope'', state', _) <- processStatements body False ast scope' state []
+                  return (scope'', state', warnings')
+
+              -- checking assertion at the end of loop iteration
+              case tryGetVar name scope'' of
+                Just z3var -> do
+                  z3ast     <- Z3.mkNot =<< Z3.mkEq z3var val
+                  validated <- validate (Local var pos) z3ast
+                  case validated of
+                    Left err    ->
+                      unroll (bound-1) forward body pos var val m scope'' state' (err:warnings')
+                    Right True  ->
+                      unroll (bound-1) forward body pos var val m scope'' state' warnings'
+                    Right False ->
+                      error "loop not validating not implemented yet"
+
+                _ ->
+                  return (scope, False, ("Loop variable " ++ show name ++ " has been invalidated") : warnings)
+
+
+    processFunctionCall :: Vars -> [AArg] -> ProcDecl -> Int -> [String] -> Z3 (Vars, Stmt, [String])
+    processFunctionCall scope aargs p state warnings =  do
       p' <- renameProc Map.empty p state
       case p' of
         (vtable, ProcDecl _ fargs' body' _, _) -> do
           scope'  <- foldM connectArgument scope (zip aargs fargs')
-          closure <- processStatements False ast scope' body'
-          scope'' <- foldM connectArgumentsBack (fst3 closure) (zip aargs fargs')
-          return $ (scope'', stmt)
+          -- closure <- processStatements False ast scope' body' 0
+          (_, scope'', _, warnings') <- processStatements body' False ast scope' 0 warnings
+          scope'' <- foldM connectArgumentsBack scope'' (zip aargs fargs')
+          return $ (scope'', stmt, warnings)
 
     findProc :: Ident -> [ProcDecl] -> ProcDecl
     findProc n procs =
@@ -457,7 +705,7 @@ processStatement doOpt ast scope stmt state =
                             , ProcDecl namep args (Ite condif' stmtsif condfi' stmtselse pos : body) posp
                             , state2)
 
-          For1 (Var t name val pv) body (Moderator (Var t2 _ val2 pv2) op expr) cond pos -> do
+          For1 (Var t name val pv) body (Moderator (Var t2 _ val2 pv2) op expr) cond b pos -> do
             (vtable', name') <- insert name state vtable
             let val'          = renameExpr vtable (fromJust val)
             tmp2             <- foldM renameStmt (vtable', ProcDecl namep [] [] posp, state + 1) body
@@ -466,12 +714,12 @@ processStatement doOpt ast scope stmt state =
                 let expr' = renameExpr vtable expr
                 let cond' = renameExpr vtable' cond
                 let mod'  = Moderator (Var t2 name' val2 pv2) op expr'
-                let for1' = For1 (Var t name' (Just val') pv) stmts mod' cond' pos
+                let for1' = For1 (Var t name' (Just val') pv) stmts mod' cond' b pos
                 return ( vtable
                         , ProcDecl namep args (for1':body) posp
                         , state1)
 
-          For2 (Var t name val pv) (Moderator (Var t2 _ val2 pv2) op expr) body cond pos -> do
+          For2 (Var t name val pv) (Moderator (Var t2 _ val2 pv2) op expr) body cond b pos -> do
             (vtable', name') <- insert name state vtable
             let val'          = renameExpr vtable (fromJust val)
             tmp2             <- foldM renameStmt (vtable', ProcDecl namep [] [] posp, state + 1) body
@@ -480,7 +728,7 @@ processStatement doOpt ast scope stmt state =
                 let expr' = renameExpr vtable expr
                 let cond' = renameExpr vtable' cond
                 let mod'  = Moderator (Var t2 name' val2 pv2) op expr'
-                let for1' = For2 (Var t name' (Just val') pv) mod' stmts cond' pos
+                let for1' = For2 (Var t name' (Just val') pv) mod' stmts cond' b pos
                 return ( vtable
                         , ProcDecl namep args (for1':body) posp
                         , state1)
@@ -537,17 +785,23 @@ processStatement doOpt ast scope stmt state =
           newVar <- makeVar t name
           case aargs of
             VarAA (Var _ namea _ _) _ _ -> do
-              let var = getVar namea scope "While connecting actual arguments with formal - variables"
-              Z3.assert =<< Z3.mkEq newVar var
-              return $ Map.insert name newVar scope
+              case tryGetVar namea scope of
+                Just var -> do
+                  Z3.assert =<< Z3.mkEq newVar var
+                  return $ Map.insert name newVar scope
+
+                _ -> return scope
 
         ArrFA t name sz _ -> do
           newVar <- makeArr t name
           case aargs of
             VarAA (Var _ namea _ _) _ _ -> do
-              let var = getVar namea scope "While connecting actual arguments with formal - arrays"
-              Z3.assert =<< Z3.mkEq newVar var
-              return $ Map.insert name newVar scope
+              case tryGetVar namea scope of
+                Just var -> do
+                  Z3.assert =<< Z3.mkEq newVar var
+                  return $ Map.insert name newVar scope
+
+                _ -> return scope
 
         ConstFA t name _ -> do
           newVar <- makeVar t name
@@ -562,39 +816,46 @@ processStatement doOpt ast scope stmt state =
     connectArgumentsBack :: Vars -> (AArg, FArg) -> Z3 Vars
     connectArgumentsBack scope (VarAA (Var _ name1 _ _) Nothing _, VarFA t name _) = do
       newVar <- makeVar t name1
-      let oldVar = getVar name scope "While connecting variable back"
-      Z3.assert =<< Z3.mkEq newVar oldVar
-      return $ Map.insert name1 newVar scope
-      
-      
-      
-      
-      --  do
-      -- let name'  = getVar name scope "While connecting var back"
-      -- let name1' = getVar name1 scope "While connecting var back"
-      -- Z3.assert =<< Z3.mkEq name' name1'
+      case tryGetVar name scope of
+        Just oldVar -> do
+          Z3.assert =<< Z3.mkEq newVar oldVar
+          return $ Map.insert name1 newVar scope
+
+        _ -> return scope
 
     connectArgumentsBack scope (VarAA (Arr _ name1 _ _ _) (Just idxs) _, ArrFA t name _ _) = do
       newVar <- makeArr t name1
-      let oldVar = getVar name scope "While connecting array back"
-      Z3.assert =<< Z3.mkEq newVar oldVar
-      return $ Map.insert name1 newVar scope
-      
-      --  do
-      -- let name'  = getVar name scope "While connecting array back"
-      -- let name1' = getVar name1 scope "While connecting array back"
-      -- Z3.assert =<< Z3.mkEq name' name1'
+      case tryGetVar name scope of
+        Just oldVar -> do
+          Z3.assert =<< Z3.mkEq newVar oldVar
+          return $ Map.insert name1 newVar scope
+
+        _ -> return scope
 
     connectArgumentsBack scope _ = return scope
 
 
-processStatements :: Bool -> [ProcDecl] -> Vars -> [Stmt] -> Z3 (Vars, [Stmt], Int)
-processStatements doOpt ast scope body = foldM (f ast) (scope, [], 0) body
+processStatements :: [Stmt] -> Bool -> [ProcDecl] -> Vars -> Int -> [String] -> Z3 ([Stmt], Vars, Int, [String])
+processStatements body doOpt ast scope state warnings =
+  foldM (f ast) ([], scope, state, warnings) body
+    >>= (\(body', scope', state', warnings') -> return (reverse body', scope', state', warnings'))
   where
-    f :: [ProcDecl] -> (Vars, [Stmt], Int) -> Stmt -> Z3 (Vars, [Stmt], Int)
-    f ast (scope, acc, state) stmt = do
-      res <- processStatement doOpt ast scope stmt state
-      return (fst res, acc ++ [snd res], state + 1)
+    f :: [ProcDecl] -> ([Stmt], Vars, Int, [String]) -> Stmt -> Z3 ([Stmt], Vars, Int, [String])
+    f ast (acc, scope, state, warnings) stmt = do
+      (scope', stmt', warnings') <- processStatement doOpt ast scope stmt state warnings
+      return (stmt':acc, scope', state + 1, warnings')
+      -- return (scope', stmt' : acc, state + 1, warnings')
+
+
+-- processStatement :: Bool -> [ProcDecl] -> Vars -> Stmt -> Int -> [String] -> Z3 (Vars, Stmt, [String])
+
+-- processStatements :: Bool -> [ProcDecl] -> Vars -> [Stmt] -> Int -> Z3 (Vars, [Stmt], Int, [String])
+-- processStatements doOpt ast scope body state = foldM (f ast) (scope, [], state) body
+--   where
+--     f :: [ProcDecl] -> (Vars, [Stmt], Int) -> Stmt -> Z3 (Vars, [Stmt], Int)
+--     f ast (scope, acc, state) stmt = do
+--       res <- processStatement doOpt ast scope stmt state
+--       return (fst res, acc ++ [snd res], state + 1, [])
 
 
 processArgs :: [FArg] -> Vars -> Z3 Vars
@@ -614,11 +875,15 @@ processArgs args scope = foldM f scope args
           return $ Map.insert n newVar vars
 
 
-processProcedure :: [ProcDecl] -> Vars  -> ProcDecl -> Z3 ProcDecl
-processProcedure ast scope (ProcDecl name args body pos) = do
+processProcedure :: [ProcDecl] -> Vars -> ([ProcDecl], [String]) -> ProcDecl -> Z3 ([ProcDecl], [String])
+processProcedure ast scope (accp, warnings) (ProcDecl name args body pos) = do
+  -- Make sure assertions created for one procedure does not influence another
+  Z3.push
   initialScope <- processArgs args scope
-  body' <- processStatements True ast initialScope body
-  return $ ProcDecl name args (snd3 body') pos
+  -- (_, body', _, newWarnings) <- processStatements True ast initialScope body 0 warnings
+  (body', _, _, newWarnings) <- processStatements body True ast initialScope 0 warnings
+  Z3.pop 1
+  return (ProcDecl name args body' pos : accp, newWarnings)
 
 
 processMainStore :: ProcDecl -> Z3 Vars
@@ -640,10 +905,10 @@ processMainStore (ProcDecl name _ body _) =
 
 
 -- First decl in decls is the main function.
-processProgram :: Program -> Z3 Program
+processProgram :: Program -> Z3 (Program, [String])
 processProgram (Program decls) = do
   initialScope <- processMainStore $ head decls
-  decls' <- mapM (processProcedure (tail decls) initialScope) decls
-  return $ Program decls'
+  (decls', warnings) <- foldM (processProcedure (tail decls) initialScope) ([], []) decls
+  return (Program $ reverse decls', reverse warnings)
 
 -- show <$> evalZ3 (liftM snd (processStatements [] Map.empty (parseStatements "local int a = 2\na += 1\ndealloc int a = 3")))