+++ CHANGELOG.md
@@ -14,12 +14,19 @@ https://ucph-ku.zoom.us/j/63343496731?pwd=Z3RiWUJ2ZmwvQ3haMTJmMmpkN1RCdz09
   - [ ] nD
 * Import and rewrite optimizer to work in new and fancy way
   - [x] Make a global store used for every procedure
-  - [ ] Implement call/uncall
+  - [x] Implement call/uncall
     - [x] Renaming to unique names
-    - [ ] Implementing a AST reverser for uncall, and other stuff
+    - [x] Implementing a AST reverser for uncall, and other stuff
   - [ ] Implement for loops
       - [ ] Invariants
-      - [ ] Constants
+        - [ ] Make unrollable invariant loops invalidate variables modified within that is not part of invariant
+      - [x] Constants
+        - [x] Make unrollable for loops invalidate variables modified within!
+      - [ ] Size n should also be seen as constant, and not variable
+        * Need some virtual table mapping array names to size
+      - [ ] Maybe allow modification of increment value within body?
+        * Either do not allow unrolling, or use extensive analysis to determine loop bound
+          * Should definately just start with the first
 * Write translator from AST to C++
 * Write playground website to use
 * profit